"Shark Bait" Zombie Pigmen and Elemental fight
Because players kept killing the pigmens in our public exp gold farm, we have added a custom advancement that will come up in chat if a player kills a zombie pigman named "Shark Bait" (or just shoots it off to its death). This message is broadcast to everyone, and is logged in console. 

~230 commands in total added for for the Elemental fight event
Keith - er, sorry, Elemental as he goes by now is an old classmate of Cao from the Mage School of Idris. He was obsessed with elemental magic, so much so that he stole some scrolls and was expelled from the academy. After we decimated Helios he took notice and has decided to attack.